Finish separating integration tests into separate files by test type
