+++ kernel/tests/s04_user_runtime_tests.rs
@@ -9,6 +9,16 @@
 //! execution taking into account the dependency chain:
 //! * `s04_*`: User-space runtimes
 
+use rexpect::errors::*;
+use rexpect::process::signal::SIGTERM;
+use rexpect::process::wait::WaitStatus;
+
+use crate::common::builder::{BuildArgs, Machine};
+use crate::common::helpers::{
+    setup_network, spawn_dhcpd, spawn_nrk, spawn_ping, spawn_receiver, DHCP_ACK_MATCH,
+};
+use crate::common::runner_args::{check_for_successful_exit, wait_for_sigterm, RunnerArgs};
+
 /// Tests the lineup scheduler multi-core ability.
 ///
 /// Makes sure we can request cores and spawn threads on said cores.
