+++ kernel/tests/s05_user_app_tests.rs
@@ -9,6 +9,15 @@
 //! execution taking into account the dependency chain:
 //! * `s05_*`: User-space applications
 
+use rexpect::errors::*;
+use rexpect::process::signal::SIGTERM;
+use rexpect::process::wait::WaitStatus;
+
+use crate::common::builder::BuildArgs;
+use crate::common::helpers::{setup_network, spawn_dhcpd, spawn_nc, spawn_nrk, DHCP_ACK_MATCH};
+use crate::common::redis::{REDIS_BENCHMARK, REDIS_PORT, REDIS_START_MATCH};
+use crate::common::runner_args::{wait_for_sigterm, RunnerArgs};
+
 /// Tests that user-space application redis is functional
 /// by spawing it and connecting to it from the network.
 ///
