+++ kernel/tests/s07_rackscale_tests.rs
@@ -9,6 +9,15 @@
 //! execution taking into account the dependency chain:
 //! * `s07_*`: Rackscale (distributed) tests
 
+use rexpect::errors::*;
+use rexpect::process::wait::WaitStatus;
+
+use crate::common::builder::BuildArgs;
+use crate::common::helpers::{
+    setup_network, setup_shmem, spawn_dcm, spawn_nrk, SHMEM_PATH, SHMEM_SIZE,
+};
+use crate::common::runner_args::{check_for_successful_exit, RunnerArgs};
+
 #[cfg(not(feature = "baremetal"))]
 #[test]
 fn s07_rackscale_phys_alloc_test() {
