+++ kernel/tests/s08_rackscale_benchmarks.rs
@@ -9,6 +9,19 @@
 //! execution taking into account the dependency chain:
 //! * `s08_*`: Rackscale (distributed) benchmarks
 
+use std::fs::OpenOptions;
+use std::io::Write;
+use std::path::Path;
+
+use rexpect::errors::*;
+use rexpect::process::wait::WaitStatus;
+
+use crate::common::builder::BuildArgs;
+use crate::common::helpers::{
+    setup_network, setup_shmem, spawn_dcm, spawn_nrk, SHMEM_PATH, SHMEM_SIZE,
+};
+use crate::common::runner_args::{check_for_successful_exit, RunnerArgs};
+
 #[test]
 #[cfg(not(feature = "baremetal"))]
 fn s08_rackscale_shmem_fxmark_benchmark() {
